feat(recipe detail page): improve layout of recipe detail page on large screens and refactor
Turned the recipe detail page into a grid layout for tablets+, and moved all components out to their
own files
