+++ src/features/recipes/components/recipe-details/Nutrition.jsx
@@ -1,0 +1,9 @@
+import PropTypes from "prop-types";
+
+export const Nutrition = () => {
+  return <p>Nutrition Information Here</p>;
+};
+
+Nutrition.propTypes = {
+  nutrition: PropTypes.array,
+};
