+++ src/features/recipes/components/recipe-details/Tips.jsx
@@ -1,0 +1,9 @@
+import PropTypes from "prop-types";
+
+export const Tips = () => {
+  return <p>Tips Here</p>;
+};
+
+Tips.propTypes = {
+  tips: PropTypes.array,
+};
